+++ src/main/java/com/dev/geunsns/apps/post/repository/PostLikeRepository.java
@@ -13,5 +13,5 @@ public interface PostLikeRepository extends JpaRepository<PostLikeEntity, Long> 
 
     Optional<PostLikeEntity> findByUserAndPost(UserEntity user, PostEntity post);
 
-    Long countByPost(PostEntity post);
+    Integer countByPost(PostEntity post);
 }
